first working for save and query(only empId)
What changes you have made?
added src/beverge.js
which is handling all the operations

Why this change is necessary?
this is a temprory change to just get it work first version
for getting it work for date query thinking to change storage pattern
from next commits it is going to be minor change in storing and accesing

What is the impact of the change?
this change will let it work for save and query(only empId) for only
valid inputs