added date query
What changes you have made?
added date and date and empId together query

Why this change is necessary?
to get transactions according to date

What is the impact of the change?
date query will not work from main because it is not connected to main
yet
but previous queries and save will work